// src/app/models/person.ts
export class Person {
    id: number;
    number_id: string;
    name: string;
    date_bird: Date;
    name_father: string;
    name_mother: string;
    marital_status: string;
    date_marriage: Date;
    name_partner: string;
    register: Date;
    gender: string;
    number_children: string;
    registration_status: string;
}